fix: Resolve SwiftLint violations in test files
- Split BottleSettingsTests into two files to fix Type Body Length and
  File Length violations
- Create EnvironmentVariablesTests.swift for environment variable tests
- Rename version15_3/version15_4 variables to sequoia153/sequoia154 to
  fix Identifier Name violations
